Initial File Upload
The original web design fragment I mentioned on the Discord chat on February 21, 2019. For the fun of it, I added the SVG file which is used as the icon in the page, and also the zipped Intaglio document. (It's a package, which is a directory on Mac OS X which is treated as a single file by the Finder.)